Add abstract settings class
